feat: Add AuthRequestModel and SignInDTO classes
This commit adds the `AuthRequestModel` class to the `auth/model` directory. This class extends the `Request` class from the `express` module and includes a `user` property of type `User`.

It also adds the `SignInDTO` class to the `auth/dto` directory. This class includes validation decorators from the `class-validator` library for the `email` and `password` properties.

These changes are necessary to support authentication functionality in the application.